requirements.txt
google-generativeai
python-dotenv
streamlit
pillow
